pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aweta-app"
version = "0.1.0"
description = "AWETA simulator/PLC I/O app"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "AWETA" }]
dependencies = [
  "python-snap7>=1.3.0",
]

[project.optional-dependencies]
dev = [
  "pyinstaller>=6.10",
]

[tool.setuptools]
zip-safe = false

[tool.setuptools.packages.find]
include = ["TIA_Db", "aweta"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md"]


